Refactor display options controls to uiBackgroundDisplayOptions
